+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v21.5.0](https://github.com/lexicalunit/spellbot/releases/tag/v21.5.0) - 2026-07-15
+
 ### Added
 
 - Added tests for i18n configuration files.